+++ CHANGELOG.md
@@ -9,6 +9,10 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+### Added
+
+- Add `@cacheControl` directive https://github.com/nuwave/lighthouse/pull/2132
+
 ## v5.48.2
 
 ### Fixed